Apply deterministic virtual_op numbering rule

+++ libraries/chain/include/graphene/chain/database.hpp
@@ -553,7 +553,7 @@ namespace graphene { namespace chain {
          uint32_t                          _current_block_num    = 0;
          uint16_t                          _current_trx_in_block = 0;
          uint16_t                          _current_op_in_trx    = 0;
-         uint16_t                          _current_virtual_op   = 0;
+         uint32_t                          _current_virtual_op   = 0;
 
          vector<uint64_t>                  _vote_tally_buffer;
          vector<uint64_t>                  _witness_count_histogram_buffer;
